requirements.txt
openai==1.23.2
python-dotenv
huggingface-hub==0.22.2
streamlit==1.33.0
langchain_openai==0.1.3
langchain-community==0.0.32
beautifulsoup4==4.12.3
pypdf==4.2.0
langchain==0.1.16
# chromadb==0.3.22
faiss-cpu==1.8.0
